Add scalac strict_dep error regexes and update logic to compare partial classnames to target closure's classes (#5093)
### Problem
There are some class not found error messages that are not covered by the included regexes. Additionally, the logic in the jvm dependency analyzer can't compare partial classnames to the classes in the target's closure. 

### Solution
Add regexes for missing class not found errors so missing-deps-suggest can make suggestions, and update the jvm dependency analyzer to ensure the partial classnames are compared to the classes in the target's closure.

Note: There was discussion of adding support for import renaming but I tested this behavior manually and it worked fine without any additional changes. I suggest possibly delaying explicit support for it.

### Results
When the error message only includes a partial classname like `http.HttpServer`
Before:
```
11:00:35 00:10         [missing-deps-suggest]

                       Found the following deps from target's transitive dependencies that provide the missing classes:
                         com.example.http.HttpServer: example/http/src/test/scala:scala

                       If the above information is correct, please add the following to the dependencies of (example/failing/target:scala):
                         'example/http/src/test/scala:scala',

                   compile(example/failing/target:scala) failed: Zinc compile failed.
FAILURE: Compilation failure: Failed jobs: compile(example/failing/target:scala)


               Waiting for background workers to finish.
11:00:41 00:16   [complete]
               FAILURE
```
After:
```
11:03:47 00:37         [missing-deps-suggest]

                       Found the following deps from target's transitive dependencies that provide the missing classes:
                         com.example.http.HttpServer: example/http/src/test/scala:scala
                         http.HttpServer: example/http/src/test/scala:scala

                       If the above information is correct, please add the following to the dependencies of (example/failing/target:scala):
                         'example/http/src/test/scala:scala',

                   compile(example/failing/target:scala) failed: Zinc compile failed.
FAILURE: Compilation failure: Failed jobs: compile(example/failing/target:scala)


               Waiting for background workers to finish.
11:03:54 00:44   [complete]
               FAILURE
```

This change causes a regression in how long missing-deps-suggest takes. Missing-deps-suggest is ~27% of the time it take to complete a cold compile and only runs once an error is detected. With this change, missing-deps-suggest is ~29% of the compile time. This is a 2-3 second difference on average.