Introduce a Digest type
This is equivalent to the Bazel Remote Execution Digest protobuf, but
without the overhead and awkward API of a protobuf.

Passing around these tuples of Fingerprint and size is a little
annoying.

Also, give it an Into implementation for Digest, because that's a thing
we do a lot with them.

This is also going to be a product type in the graph, so having a type
for it is useful.