Snapshots can be captured as store-backed Directories as well as tar files.
A follow-up PR will remove the tar file implementation, and with it, a lot of code.

Moves Directory traversal logic from fs_util into fs

Introduces a GetFileDigest trait which can be implemented both by Context (for memoized graph operations) and in fs_util and tests (for one-off file reading).