Remove inlining in favor of executing directly (#5095)
### Problem

Memoization is currently enabled when Node is implemented. When memoization isn't required, Node wouldn't need to be implemented so removing is_inline from the signature of Node trait and executing directly.

### Solution

Don't implement the `Node` trait for nodes that are currently inline, and remove them from relevant `Node`-related enums.