Extract resettable pool logic for reuse (#5110)
### Problem

The resettable `CpuPool` in `PosixFS` should be consumed by `Store` in order to parallelize requests to `LMDB`, but currently it is tightly coupled to its parent struct.

### Solution

Extract `ResettablePool`, and prepare for `Core` to pass it to both `PosixFS` and `Store`. Also, take this opportunity to improve the API for accessing the pool.

### Result

Fixes #5108.