Async Store (#5117)
### Problem

`Store` executes IO, but was doing so synchronously on the main thread, which might block other operations.

### Solution

Give `Store` a reference to a `ResettablePool`, and execute all database operations on the pool.

It was also necessary to apply the "`Arc<Inner*>` pattern" (maybe it has a name?) to allow the `Store` to be cloned into the pool... this pattern seems much cleaner than the "`impl Trait for Arc<Struct>` pattern" that we've used with the `VFS` implementations... should consider switching those over at some point.

Finally, I added `Store::load_bytes_with` to support copying directly from LMDB buffers into a destination to avoid copying, and execute the copy on the IO pool.

### Result

`save_directory` and `materialize_directory` in `fs_util`, and `DigestFile` in `nodes` will execute concurrently. Helps to unblock #5105 and #5106.