Remove requirement for root build file in `changed` (#5134)
### Problem

The `changed` options as implemented in the v2 engine use the `AscendantAddresses` spec type to find "all addresses above a directory". As described in #4833 and #4904, this currently introduces a requirement that at least one `BUILD` file exists between a changed file and the root of the repo.

### Solution

Remove the requirement for a non-empty set of `AddressFamily` instances when expanding `AscendantAddresses`, and add a covering integration test.

### Result

`AscendantAddress` instances that don't encounter `BUILD` files will not cause errors.

Fixes #4833 and #4904.