Peek Target using Buildozer
This adds the ability to peek a BUILD file for a named target, printing the target name if it is found.Additionally, peek allows a line_number flag which will print the starting line of the named target if it is found in the BUILD file.

Also, this adds a suppress warnings feature to Buildozer.

Example:
$ ./pants peek
--line_number=true path/to/directory:target_name

To Verify:
$ ./pants test
contrib/buildrefactor/tests/python/pants_test/contrib/buildrefactor:peek_integration

This addresses:
https://github.com/pantsbuild/pants/issues/4861 Please enter the commit message for your changes. Lines starting