Build suffixed pants wheels for non-releases (#5118)
### Problem

As described on #4896, the "binary builder" shards in travis currently overwrite the previous artifacts for a SHA, meaning that it is possible for `release.sh` to race with additional commits landing in master.

Relatedly, because we build artifacts for all shas and not just tags (and should continue to do so in order to solve #4896), it's not possible to differentiate the `whl`s that were built for an actual release from those built for any old dev sha. 

### Solution

1. Build version-suffixed `whl`s for all travis builds.
    * This will allow consumers to build a pex for any SHA that has been built in travis unambiguously, while avoiding cache collisions due to re-used version numbers.
2. SHA-prefix the artifacts built by the "binary builder" shards to avoid problematic collisions between published artifacts.
3. "Re-version" whls that we release to PyPI (via #5145).

### Result

This change makes releases less racy, and begins to unblock building pexes from any SHA with #4896 in future PRs.