[pantsd] Repair stdio truncation. (#5156)
### Problem

The introduction of `os.pipe()` usage for daemon stdio streams introduced 3 new threads (stdin, stdout, stderr) involved in every pantsd-runner run. Because the writers are running in python threads, it's possible (and likely, on my machine) for the main thread to write an `EXIT` chunk before the `NailgunStreamWriter` threads are able to wake up and write their payloads to the socket.

### Solution

Construct a finalizing function and pass this to the `DaemonExiter` to invoke before sending any exit chunks. This finalizer will stop and join the writer threads giving the writers a chance to emit any remaining output before termination.

### Result

Output truncation is no longer reproducible.