Fix RunTracker re-initialization with pantsd (#5211)
### Problem

As described in #5140, `./pants server` was rendering spurious broken runs in the presence of `pantsd`.

This was due to `RunTracker.__init__` triggering sideeffects, including the creation of the `run_info_dir` and `RunInfo` (which itself has sideeffects). The act of calling `RunTracker.global_instance()` was causing `__init__` to run once... it would then be called again after `Subsystem.reset`. The result was a partial `info` file that was never finalized, and thus caused `./pants server` to crash at consumption time.

### Solution

I initially considered only removing the call to [RunTracker.global_instance().reset()](https://github.com/pantsbuild/pants/blob/master/src/python/pants/init/util.py#L20-L22), which avoids the underlying issue, but doing that alone seemed likely to be fragile in the case of additional accesses to the `RunTracker` being added in the future.

So I additionally changed the lifecycles of `RunTracker`/`Reporting` to be more explicitly intertwined, and to make the creation of the `RunInfo` part of those lifecycles. The two classes already had implicit dependencies via direct access to the `run_info` instance... more work to do there.

### Result

One (valid) `RunTracker` run_id is created per `pants` or `pantsd` run. Fixes #5140.