TargetRoots always requires options
There are a surprising number of places where options parsing will
implicitly be done if you omit an optional argument.

This is one of a series of PRs which will push options parsing out to
happen where we actually have a reference to the arguments we're
parsing, rather than to implicitly happen wherever we may happen to need
options.