Log and increase pantsd startup timeout (#5231)
### Problem

As described on #5227, the 10 second startup time for the daemon is occasionally too short on a busy machine.

### Solution

Since the startup time of the daemon itself should not be proportional to the size of the repo, but just how busy the machine is, set the timeout much higher rather than making it configurable. Additionally, log why we are waiting (if we do end up waiting) in order to explain the potentially much longer delay.

### Result

Busy machines should not see spurious failures to start the daemon. Fixes #5227.