Swap stdio file descriptors at the os level (#5247)
### Problem

As explained in comments on #5221, due to how we were replacing the file descriptors that backed `sys.{stdin,stdout,stderr}`, it was possible for `subprocess` calls to behave differently in the daemon vs outside the daemon. In short: `subprocess` contains assumptions that the file descriptors behind the stdio handles have file descriptors `0, 1, 2`. 

### Solution

Update `stdio_as` to make os-level replacements of file descriptors, in addition to the replacements of the python-level handles. As explained in the docstring there, it is impossible to tell whether the replaced file descriptors in `0, 1, 2` have been leaked/aliased for other usages, so the caller needs to be careful that the handles they are replacing are safe to replace.

This is accomplished in the context of the `pantsd` process by calling `stdio_as` immediately after we close the stdio file handles, guaranteeing that they are reusable. We replace them with open (but useless: pointed to `/dev/null`) handles to reserve the `0, 1, 2` file descriptors for usage in forked runners.

Finally, in the `pantsd-runner` process, the reserved `0, 1, 2` slots are replaced (again at the os fd level) with pipes that are copied to the nailgun socket.

### Result

#5221 is fixed, and further recurrences of that issue are prevented since `subprocess` calls act exactly as they would outside the context of the daemon.