Goal-level --transitive and --skip options for lint/fmt tasks.
This change introduces a lightweight mechanism for registering
options at the goal level. This is useful for
A) Registering recursive options on all tasks in a goal in a single place.
B) Being able to set the option value at the goal level for all tasks.

This change then uses this mechanism to provide --transitive and --skip
to all lint and fmt tasks.  To do so it introduces mixins that handle
computing the target set based on the values of those options.

This involved adding a new method to the Task base class: get_targets().
Tasks are obviously not required to use this, but it would be nice to
retrofit existing tasks at some point to do so.