A lightweight mechanism for registering options at the goal level.
This is useful for:
A) Registering recursive options on all tasks in a goal in a single place
   (although this could also have been achieved with a shared base class).
B) Being able to set the option value at the goal level for all tasks.

This change also demonstrates use of this mechanism for providing
--transitive and --skip options to all lint and fmt tasks.
This requires not only shared registration (via the new mechanism)
but also shared target set computation functionality. The latter
is achieved via Task mixins.

This involved adding a get_targets() method to the Task base class.
This method provides a useful point at which to isolate the computation of
the target set.

Tasks are obviously not required to use get_targets(), but it would be nice to
retrofit existing tasks at some point to do so.

Note that this change doesn't actually apply the new mechanism to
any lint or fmt tasks. A followup change will do so.