Deprecate the v1 engine option. (#5338)
### Problem

The v1 engine is on the way out, but we have not yet deprecated enabling it. Part of the reason for waiting is that we wanted the daemon to be in a usable state before requiring usage of v2: it now is.

### Solution

Deprecate enabling the v1 engine. Will be cherry-picked to the `1.4.x` branch.