Reenable lighter contrib sanity checks (#5340)
### Problem

In short: #4992 skipped running the `pkg_<name>_install_test` function for contrib modules because some of those modules defined expensive checks, and because all contrib tests are run on dedicated Contrib shards already. But that, unfortunately, meant that we didn't even sanity check that contrib modules could be loaded as plugins.

### Solution

Simplify all `pkg_<name>_install_test` functions to be simple checks that tasks are appropriately registered, and re-enable running them on the binary builder shard. Also, fix some registration issues that slipped through with the confluence module due to this gap.

### Result

#5339 is fixed.