Store and remote::ByteStore use Digests not Fingerprints
local::ByteStore still uses Fingerprints

This allows us to use proper identifiers when interacting with a CAS,
which are specified to include length.

It also means that Store's interface better matches the Graph's
interfaces.

This is a breaking change for fs_util, as it no has an extra required
argument wherever before it just took a fingerprint.