python2: do not resolve requirements if no python targets in targets closure (#5361)
### Problem

If Python targets depend on `resources` targets, the resolve_requirements task will be invoked even if the current build does not have any Python targets in the targets to be processed. The resolve_requirements task should not do any work in that case.

### Solution

Only resolve requirements if there are requirements actually needing resolution.

### Result

Doesn't do the problematic work.