Garbage collect Store entries (#5345)
A caller may lease a blob in the Store which will ensure it isn't
garbage collected for a time period.

A caller may shrink the store to n bytes (ignoring lmdb overhead), which
will walk a list of blobs which are unleased, and delete them in the
order which their most recent lease expired, until the store is small
enough.

The pantsd python code occasionally triggers lease extensions of
everything currently present in the Graph, and occasionally triggers
garbage collection.

This also removes the `ByteStore` trait which existed on local and
remote `ByteStore`s to make testing easier. They no longer have the
same API, because the local one cares about leases (to be fair, they
didn't have the same API before either, but I pretended they did for
testing convenience).