Update our JVM-related config and documentation.
Our docs referenced JVM 6 and 7, both of which were
end-of-lifed long ago. This change updates those to
reference 8 and 9, so the documentation doesn't look
outdated.

Also pins the JVM version we build and run on in the
Pants repo to 1.8, and removes uses of MaxPermSize,
which does nothing in JVM 8+ and generates a warning.

Note that this change does not modify tests for language
versioning, e.g., for cross-compiling to a specific Java
language version. Java 6 (the language version that, e.g.,
doesn't support the diamond operator) is still a valid
concept, even though JVM 6 is long dead.