Repoint the 'current' symlink even for valid VTs.
Previously we only repointed it for invalid ones, which meant
that the "current" link could be stale if you had two valid versions
of a target and were switching between them (e.g., by reverting
a change).

This change also renames results_dir_path() to _results_dir_path(),
to emphasize that it's not for outside consumption.

It also removes a superfluous arg to a test helper method, fixes
a bug in a test where we used assertTrue instead of assertEquals,
and removes a superfluous 'or False' from a boolean condition.