No-op ivy resolve is ~100ms cheaper
Right now, creating an executor can take as much as 150ms. Instead, only
create the executor if it's actually needed.

This is a bit of an abstraction leak, but 150ms is non-trivial (and we
have a *lot* of this kind of 150ms kicking about, which add up to
several seconds).