Python distribution task for user-defined setup.py + integration with ./pants {run/binary/test} (#5141)
Problem

There is no way to consume a python_library with c_extensions. Relates to #4740.

Solution

Create a new target, python_distribution, that specifies only a distribution name that maps to the directory the BUILD file lives in (a distribution's project directory). Pants will treat the directory that the python_distribution target is defined in aa a distribution project directory (containing a setup.py and sources) and run the pex command on that directory to build an inline wheel using the setup.py. The ./pants binary and ./pants run tasks will dump this distribution into the binary pex for consumption (or requirements pex for the ./pants run case). The setup.py takes care of linking the package sources and c/cpp code together as setuptools Extensions. We've run a similar example by our customer to validate that this meets their needs.

Result

Running ./pants binary or ./pants run on a python_binary target that depends on a python distribution will package the distribution as a wheel and include it in the created pex binary or requirements pex (for ./pants run) for consumption. This example has simple C code to demonstrate the ability to call C extensions from a package built by using this python_distribution target.