Fetch the buildozer binary using a subsystem.
This is part of our transition to registering all binary tools
explicitly, instead of letting code access them arbitrarily.

This required some refactoring of the buildrefactor tasks.

This change also fixes a goof in the new protoc subsystem.

It also adds a way to specify a tool name different from
the options_scope, which is necessary here for legacy reasons.