[coursier/m2-coords] update coursier json parsing; use maven's coords
This is a multi-tool change. The coursier side is here: https://github.com/coursier/coursier/pull/782

This fixes the issue with coursier resolution where transitive dependencies that have classifiers can't be differentiated. Ie, if you have a classifier on a dependency, you get all the jars even if they have a different classifier.

Additionally this updates the m2 coordinate string representation to bring it inline with maven's coordinate language. This is because I've chosen that language to use in the coursier change as well