Snapshot is backed by LMDB not tar files
This is much more incremental, and will help us to support remote
execution (where Directory protos are the lingua franca).

There are still several significant things to clean up here:
 * Some IsolatedProcess tests are disabled
 * We need to work out how to do input/output conversions for
   IsolatedProcess for real
 * Process execution is hard-coded to be local, synchronous, and a
   little panic-y

but as this is already a very large change, and the process execution
pieces aren't actually used in production codepaths, I'd like those to
be follow-ups.