Don't reinstall plugin wheels on every invocation.
The wheel install dir path completely qualifies it. So, assuming
it was moved atomically to that path, we can take existence of
the directory to imply that it contains the installed wheel.

This saves ~2 seconds (70%) of the no-op ./pants invocation time
in a repo with just a handful of plugins.