Move Key interning to rust (#5455)
### Problem

Key interning was implemented on the python side of the FFI boundary, which meant spending a longer time on the python side than was strictly necessary in the `key_for` and `val_for` externs.

### Solution

Move the Key intern map and the `key_for`/`val_for` functions to rust such that they can spend less time with the GIL acquired, and avoid a few clones.

### Result

`./pants list ::` is about ~8% faster in a large repository. No change in memory usage.