New-style BinaryTool Subsystems for isort and go distribution. (#5523)
Gets rid of the indirection via a Factory. It's easier to just
make the distribution itself the subsystem, since the base class
takes care of basically everything the factory previously did.

Also improved the fetching code to use a proper tmpdir instead
of a hardcoded name that might cause contention.