Add ability for pants to call coursier with the new url attribute (#5527)
### Problem
We need to fetch a jar with a url using coursier but previously this was not something coursier could do. With the [update to coursier](https://github.com/coursier/coursier/commit/3e4a65d5ee66f043c2467972bd6e29f48b570715) that adds this functionality, we need to update pants to use it. 

### Solution
All that is required to integrate the coursier changes is to check if the url is present for the jar dep and call coursier with that jar dep. An additional change was made to adjust the path of the jar file if it is a local url (coursier doesn't symlink local files to it's cache).

### Result
After updating the coursier jar link, it works! I am able to fetch the jar file for a dependency that has defined the url.