Engine is a workspace
Now that we're using a cdylib, this can be enabled.

This allows rust-standard ways of operating across a whole project, like
running `cargo test --all`.

It also improves compile time. Using a workspace means that a single
`Cargo.lock` file is used for the project, and a single `target`
directory is used for outputs. This in turn means that if building
multiple components (e.g. `engine` and `fs_util`), the intermediate
artifacts are actually shared across components, rather than
re-compiled.