Shard lmdb by top 4 bits of fingerprint
Currently, we can only write files from a single thread at a time, which
is a significant bottleneck in large repos.

Garbage collection will be slightly slower, as we can no longer share
one txn for all garbage collection operations, but this shouldn't be the
end of the world. If it turns out to be a problem, we can optimise it.