Only shift once, not twice (#5552)
cb72c2aa5f188e26ef93a8fb2a18524daa75a1ac removed the CORE_ONLY which was
the second shift target.

It turns out that bash no-ops if you `shift 2` when there is only one
arg. So that's fun.

Also, `set -e` for this script. Note that it doesn't actually do
anything about the `shift` error (that would be far too useful), but
it's something we should generally be doing.

But also, we should rewrite this script in python or something, because
it's gotten insane. Filed
https://github.com/pantsbuild/pants/issues/5551 to track that.