Don't force fsync on every lmdb write transaction
This significantly improves performance on slow or contended disks.

On filesystems which preserve order of writes, on system crash this may
lead to some transactions being rolled back. This is fine because this
is just a write-once content-addressed cache. There is no risk of
corruption, just compromised durability.

On filesystems which don't preserve the order of writes, this may lead
to lmdb corruption on system crash (but in no other circumstances, such
as process crash).