Render a warning rather than failing `list` when no targets are matched (#5598)
### Problem

The `v2` engine changed the behaviour of `./pants list` (with no specs) from "list everything", to "fail with an error". But due to the deprecation/removal of the `changed` goal, this also has the effect of failing with an error for `./pants --changed=.. list` calls that result in no matched targets, which is a behaviour change from `v1`.

### Solution

Rather than raising an error, trigger a warning to stderr which naturally handles both cases by indicating that their query wasn't an error: it just didn't match anything.