Deprecate BinaryUtil as public API. (#5601)
Clients shouldn't use it directly, because we want to be able to discover which binary tools are "registered" for use via BinaryToolBase subclass subsystems.

This change also gets rid of a heavy-handed mechanism for checking if two files have the same content. The old code was convoluted and hard to reason about, and included degrees of freedom (such as setting the hasher) that weren't used in practice. It also contained a bug (it used a hashlib.sha1(), which is mutable, as a default value for a method arg).

The replacement logic uses the built-in filecmp module. This is much more straightforward, and likely faster, than computing two hashes and comparing them.