[rust/engine] add timeout for grpc
https://github.com/pantsbuild/pants/issues/5504

This is an initial movement towards adding a timeout. It still needs more work.

This patch: On responses from the server, instead of always continuing, we'll now timeout if the incomplete responses take more than the timeout. Currently in the patch, the timeout is 4 seconds, which is quite unrealistic, but it makes it easier to test, given that the timeout isn't injected yet.

Before this can land, it needs to inject the timeout duration into the client rather than having a hardcoded value. It also might make sense to investigate other future framework mechanisms for handling timing out. This patch's construction assumes that the server will be quick in its responses about incomplete operations.