Remove ExecutionEnvironmentMixin and expose path entries instead
In response to review comments, remove a lot of indirection and instead just
exposed the native toolchain as path entries, and allow consumers such as
BuildLocalPythonDistributions to perform the environment modification.

Also add a lot of explanatory remarks, some still in comments.

Also introduce `XCodeCLITools` to wrap the OSX provided toolchain uniformly.