Resolve for current platform only if resolving a local python dist with native extensions (#5618)
### Problem

If a pants.ini file specifies default platforms that include more than just the current platform, local python dist targets will fail to resolve during the resolve requirements task because they are built locally for the current platform. Furthermore, the resolve requirements task does not plumb binary target platform arguments through to the pex resolver at resolve time, so we need a mechanism to check if there are native sources (from a python dist) in play, and pass only the `"current"` platform to the pex resolver if this is the case.

### Solution

Check if a python_dist target is in play during the requirements resolution task and whether any of these targets contain native (c or cpp) sources. If they do, set the `platforms` parameter of the pex resolver to contain only `"current"`.

### Result

Pants projects that have pants.ini files which specify multiple default platforms can now build local python dist targets for the current platform only.